Arreglados errores de elementos de individual
