+++ src/main/java/acme/roles/Lecturer.java
@@ -8,7 +8,7 @@ import javax.validation.constraints.NotBlank;
 import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.URL;
 
-import acme.entities.course;
+import acme.entities.course.Course;
 import acme.framework.data.AbstractRole;
 import lombok.Getter;
 import lombok.Setter;
@@ -36,5 +36,5 @@ public class Lecturer extends AbstractRole {
 	private String				link;
 
 	@ManyToOne()
-	protected course			courses;
+	protected Course			courses;
 }
